+++ stratum/hal/lib/barefoot/bf_sde_mock.h
@@ -91,6 +91,7 @@ class BfSdeMock : public BfSdeInterface {
   MOCK_METHOD1(GetPcieCpuPort, ::util::StatusOr<int>(int device));
   MOCK_METHOD2(SetTmCpuPort, ::util::Status(int device, int port));
   MOCK_METHOD1(IsSoftwareModel, ::util::StatusOr<bool>(int device));
+  MOCK_CONST_METHOD1(GetBfChipType, std::string(int dev_id));
   MOCK_METHOD2(TxPacket, ::util::Status(int device, const std::string& packet));
   MOCK_METHOD1(StartPacketIo, ::util::Status(int device));
   MOCK_METHOD1(StopPacketIo, ::util::Status(int device));
